Prepare to productize when_all
