+++ groups/nts/ntsa/ntsa_coroutine.t.cpp
@@ -2733,517 +2733,3 @@ NTSCFG_TEST_FUNCTION(ntsa::CoroutineTest::verifySandbox)
 }  // close namespace BloombergLP
 
 #endif  // NTS_BUILD_WITH_COROUTINES
-
-// MRM
-#if 0
-
-
-
-using namespace BloombergLP;
-using namespace bsl;
-
-
-
-// ============================================================================
-//                                 TEST PLAN
-// ----------------------------------------------------------------------------
-//                                 Overview
-//                                 --------
-// TODO
-// ----------------------------------------------------------------------------
-// [ 1] CoroutineTask<>
-// [ 2] RESULT CoroutineUtil::sync_wait(CoroutineTask<RESULT>&&);
-// [ 3] CoroutineTask(CoroutineTask&&)
-// [ 4] `co_await` for CoroutineTask
-// ----------------------------------------------------------------------------
-// [ 1] BREATHING TEST
-// [ 5] USAGE EXAMPLE
-
-// ============================================================================
-//                     STANDARD BDE ASSERT TEST FUNCTION
-// ----------------------------------------------------------------------------
-
-namespace {
-
-int testStatus = 0;
-
-void aSsErT(bool condition, const char *message, int line)
-{
-    if (condition) {
-        cout << "Error " __FILE__ "(" << line << "): " << message
-             << "    (failed)" << endl;
-
-        if (0 <= testStatus && testStatus <= 100) {
-            ++testStatus;
-        }
-    }
-}
-
-}  // close unnamed namespace
-
-// ============================================================================
-//               STANDARD BDE TEST DRIVER MACRO ABBREVIATIONS
-// ----------------------------------------------------------------------------
-
-
-
-
-
-
-// ============================================================================
-//                  NEGATIVE-TEST MACRO ABBREVIATIONS
-// ----------------------------------------------------------------------------
-
-
-
-// ============================================================================
-//                        GLOBAL HELPER FUNCTIONS FOR TESTING
-// ----------------------------------------------------------------------------
-#ifdef BSLS_LIBRARYFEATURES_HAS_CPP20_BASELINE_LIBRARY
-namespace {
-
-using Util = bdxa::CoroutineUtil;
-
-
-
-// ============================================================================
-//                                TEST CASE 4
-// ----------------------------------------------------------------------------
-
-namespace test_case_4 {
-// A struct that can be on the left-hand side of a comma expression only if
-// the right-hand side has type void.
-struct EnforceVoid {
-    void operator,(auto&&) = delete;
-};
-
-// Verify that an lvalue of type `CoroutineTask<RESULT>` is not awaitable.
-template <class RESULT = void>
-void doP3()
-{
-    using Task = bdxa::CoroutineTask<RESULT>;
-    ASSERT(!requires (Task task) { task.await_ready(); });
-    ASSERT(!requires (Task task) { task.operator co_await(); });
-    ASSERT(!requires (Task task) { bdxa::operator co_await(task); });
-}
-}  // close namespace test_case_4
-
-// ============================================================================
-//                                TEST CASE 2
-// ----------------------------------------------------------------------------
-
-namespace test_case_2 {
-// Verify that `ntsa::CoroutineUtil::sync_wait` doesn't accept an lvalue of type
-// `CoroutineTask<RESULT>`.
-template <class RESULT = void>
-void doP4()
-{
-    ASSERT(!(requires (bdxa::CoroutineTask<RESULT> task) {
-                 ntsa::CoroutineUtil::sync_wait(task);
-             }));
-}
-}  // close namespace test_case_4
-
-// ============================================================================
-//                               USAGE EXAMPLE
-// ============================================================================
-
-//Usage
-//-----
-// This section illustrates intended use of this component.
-//
-//Example 1: TODO
-//- - - - - - - - - - - - - - - - - -
-// TODO
-// ```
-void usageExample()
-{
-    ASSERT(true);
-}
-// ```
-}  // close unnamed namespace
-#endif  // C++20
-
-//=============================================================================
-//                                MAIN PROGRAM
-//-----------------------------------------------------------------------------
-
-int main(int argc, char *argv[])
-{
-    int test    = argc > 1 ? bsl::atoi(argv[1]) : 0;
-    int verbose = argc > 2;
-
-    cout << "TEST " << __FILE__ << " CASE " << test << '\n';
-    switch (test) { case 0:
-      case 5: {
-        // --------------------------------------------------------------------
-        // USAGE EXAMPLE
-        //   Extracted from component header file.
-        //
-        // Concerns:
-        // 1. The usage example provided in the component header file compiles,
-        //    links, and runs as shown.
-        //
-        // Plan:
-        // 1. Incorporate usage example from header into test driver, remove
-        //    leading comment characters, and replace `assert` with `ASSERT`.
-        //    (C-1)
-        //
-        // Testing:
-        //   USAGE EXAMPLE
-        // --------------------------------------------------------------------
-
-        if (verbose) cout << "\nUSAGE EXAMPLE"
-                             "\n=============\n";
-#ifdef BSLS_LIBRARYFEATURES_HAS_CPP20_BASELINE_LIBRARY
-		usageExample();
-#else
-        cout << "Skipping usage example before C++20...\n";
-#endif
-      } break;
-      case 4: {
-        // --------------------------------------------------------------------
-        // `co_await`
-        //
-        // Concerns:
-        // 1. `co_await`ing a `CoroutineTask<T>` suspends the awaiting coroutine
-        //    C, resumes the `CoroutineTask` coroutine, and causes the latter to
-        //    resume C upon completion, at which point the value of the
-        //    `co_await` expression is the `T` that was initialized by the
-        //    `co_return` statement in the `CoroutineTask` coroutine.
-        // 2. An lvalue of `CoroutineTask` is not awaitable.
-        //
-        // Plan:
-        // 1. Create a coroutine template that returns `CoroutineTask` of the
-        //    template argument, which can be an object type, void, lvalue
-        //    reference, or rvalue reference.  Define this coroutine template
-        //    to `co_return` a pre-set value of the specified return type.
-        //    For each kind of result type, create a `CoroutineTask`, `T1`, by
-        //    instantiating the template, and another `CoroutineTask`, `T2`, that
-        //    returns the same type, referring a coroutine that `co_await`s
-        //    `T1` and `co_return` the result of the `co_await` expression.
-        //    Use `sync_wait` to check that the result of `T2` is the expected
-        //    pre-set value.  (C-1)
-        //
-        //    Note that the expected behavior is obtained only because a
-        //    `CoroutineTask` coroutine returns control to the awaiter of the
-        //    `CoroutineTask` upon completion of the coroutine.  The sequence of
-        //    events is as follows:
-        //    1. `sync_wait` awaits `T1`, resuming the coroutine;
-        //    2. The coroutine referred to by `T1` awaits `T2`, resuming the
-        //       coroutine;
-        //    3. The body of `T2` runs, and the coroutine executes the implicit
-        //       `co_await` expression at the final suspend point;
-        //    4. The `await_suspend` method returns the coroutine handle of
-        //       `T1`, causing it to be resumed;
-        //    5. `T1` finishes executing the `co_await` expression and runs to
-        //       its final suspend point.  The coroutine in `sync_wait` is then
-        //       resumed like in step 4;
-        //    6. The coroutine in `sync_wait` sets a flag and the wait
-        //       completes.
-        //
-        //    If `CoroutineTask` coroutines were programmed not to suspend at the
-        //    final suspend point, then the following (undesired) sequence of
-        //    events would occur:
-        //    1. `sync_wait` awaits `T1`, resuming the coroutine;
-        //    2. The coroutine referred to by `T1` awaits `T2`, resuming the
-        //       coroutine;
-        //    3. `T2` runs to completion and returns;
-        //    4. The `co_await T2` expression in the coroutine referred to by
-        //       `T1` returns control to the resumer of that coroutine;
-        //    5. `sync_wait` waits forever for the coroutine referred to by
-        //       `T1` to complete.
-        //
-        //    If `CoroutineTask` coroutines were programmed to always suspend at
-        //    the final suspend point (returning control to the most recent
-        //    awaiter), then the same (undesired) sequence of events occurs
-        //    except that `T2` remains suspended at its final suspend point.
-        // 2. Create a `CoroutineTask`, `T1`, referring to a coroutine that throws
-        //    an exception, and another `CoroutineTask`, `T2`, that `co_await`s
-        //    `T1` within a try-catch block and `co_return`s the caught value.
-        //    Use `sync_wait` to verify that the result of `T2` is the expected
-        //    value.  (C-1)
-        // 2. Use a requires-expression to verify that an lvalue of type
-        //    `CoroutineTask` doesn't have a callable `await_ready` method or
-        //    `operator co_await` method, and there is no free
-        //    `operator co_await` function that accepts it.  (C-2)
-        //
-        // Testing:
-        //   `co_await` for CoroutineTask
-        // --------------------------------------------------------------------
-
-        if (verbose) cout << "\n`co_await`"
-                             "\n==========\n";
-
-#ifdef BSLS_LIBRARYFEATURES_HAS_CPP20_BASELINE_LIBRARY
-        using namespace test_case_4;
-        using bdxa::CoroutineTask;
-
-        // P-1
-        {
-            // We can't put `co_await` inside `decltype`, so the only way we
-            // can check whether a `co_await` expression has type `void` is to
-            // use the comma expression trick (see definition of the
-            // `EnforceVoid` helper class).
-            auto task = []() -> CoroutineTask<bool> {
-                            EnforceVoid(), co_await coReturnValue<void>();
-                            co_return true;
-                        }();
-            ASSERT(ntsa::CoroutineUtil::sync_wait(bsl::move(task)));
-        }
-        {
-            auto task = []() -> CoroutineTask<int> {
-                            co_return co_await coReturnValue<int>();
-                        }();
-            ASSERT(getValue<int>() == ntsa::CoroutineUtil::sync_wait(bsl::move(task)));
-        }
-        {
-            auto task = []() -> CoroutineTask<int&> {
-                            co_return co_await coReturnValue<int&>();
-                        }();
-            ASSERT(&getValue<int&>() == &ntsa::SimplTaskUtil::sync_wait(bsl::move(task)));
-        }
-        {
-            auto task = []() -> CoroutineTask<int&&> {
-                            co_return co_await coReturnValue<int&&>();
-                        }();
-
-            int&& expectedRef = getValue<int&&>();
-            int&& actualRef   = ntsa::CoroutineUtil::sync_wait(bsl::move(task));
-            ASSERT(&expectedRef == &actualRef);
-        }
-
-#ifdef BDE_BUILD_TARGET_EXC
-        // P-2
-        {
-            auto task = []() -> CoroutineTask<bsl::optional<int>> {
-                            bsl::optional<int> thrownValue;
-                            try {
-                                co_await []() -> CoroutineTask<void> {
-                                             throw 2;
-                                         }();
-                            } catch (int ex) {
-                                thrownValue = ex;
-                            }
-                            co_return thrownValue;
-                        }();
-            ASSERT(2 == ntsa::CoroutineUtil::sync_wait(bsl::move(task)));
-        }
-#endif
-
-        // P-3
-        doP3();
-#else
-        cout << "Skipping test case 4 before C++20...\n";
-#endif
-      } break;
-      case 3: {
-        // --------------------------------------------------------------------
-        // MOVE CONSTRUCTION
-        //
-        // Concerns:
-        // 1. The move constructor of `CoroutineTask` creates a `CoroutineTask`
-        //    object that refers to the coroutine that the source referred to.
-        // 2. The default constructor, copy constructor, copy assignment
-        //    operator, and move assignment operator of `CoroutineTask` are
-        //    deleted or not declared.
-        //
-        // Plan:
-        // 1. Create a `CoroutineTask` coroutine and call it.  Then,
-        //    move-construct a `CoroutineTask` object from the result of the call.
-        //    Finally, use `sync_wait` to resume the coroutine through the
-        //    newly created `CoroutineTask` and verify that the expected result is
-        //    obtained.  (C-1)
-        // 2. Use the appropriate standard library traits to verify the
-        //    properties described in (C-2).  (C-2)
-        //
-        // Testing:
-        //   CoroutineTask(CoroutineTask&&);
-        // --------------------------------------------------------------------
-
-        if (verbose) cout << "\n`MOVE CONSTRUCTION`"
-                             "\n==================\n";
-
-#ifdef BSLS_LIBRARYFEATURES_HAS_CPP20_BASELINE_LIBRARY
-        using bdxa::CoroutineTask;
-
-        auto task  = []() -> CoroutineTask<int> {
-            co_return 4;
-        }();
-        auto task2 = bsl::move(task);
-        ASSERT(4 == ntsa::CoroutineUtil::sync_wait(bsl::move(task2)));
-
-        ASSERT(!bsl::is_default_constructible_v<CoroutineTask<>>);
-        ASSERT(!bsl::is_copy_constructible_v<CoroutineTask<>>);
-        ASSERT(!bsl::is_copy_assignable_v<CoroutineTask<>>);
-        ASSERT(!bsl::is_move_assignable_v<CoroutineTask<>>);
-#else
-        cout << "Skipping test case 3 before C++20...\n";
-#endif
-      } break;
-      case 2: {
-        // --------------------------------------------------------------------
-        // `CoroutineUtil::sync_wait`
-        //
-        // Concerns:
-        // 1. The `sync_wait` method waits for the `CoroutineTask` coroutine to
-        //    complete, then returns the value or propagates the exception with
-        //    which the coroutine exited.
-        // 2. The thread on which the coroutine completes can be either the
-        //    same thread or a different thread.
-        // 3. `sync_wait` doesn't accept lvalues.
-        //
-        // Plan:
-        // 1. Create a coroutine template that returns `CoroutineTask` of the
-        //    template argument, which can be an object type, void, lvalue
-        //    reference, or rvalue reference.  Define this coroutine template
-        //    to `co_return` a pre-set value of the specified return type.
-        //    Verify that in each case, the result of calling `sync_wait` is
-        //    the pre-set value of the specified type.  (C-1,2)
-        // 2. Create a `CoroutineTask` coroutine that throws an exception.  Call
-        //    `sync_wait` within a try-catch block and verify that the
-        //    expected exception was propagated from `sync_wait`.  (C-1,2)
-        // 3. Create an awaitable that will suspend the awaiter and signal a
-        //    configured thread to resume the awaiter.  Create a `CoroutineTask`
-        //    coroutine that will `co_await` such an awaitable.  Use
-        //    `sync_wait` to verify that the calling thread obtains the value
-        //    that is eventually returned by the `CoroutineTask` coroutine.
-        //    (C-2)
-        // 4. Use a requires-expression to verify that a call to `sync_wait` is
-        //    ill formed when the argument is an lvalue.  (C-3)
-        //
-        // Testing:
-        //   RESULT CoroutineUtil::sync_wait(CoroutineTask<RESULT>&&);
-        // --------------------------------------------------------------------
-
-        if (verbose) cout << "\n`sync_wait`"
-                             "\n===========\n";
-
-#ifdef BSLS_LIBRARYFEATURES_HAS_CPP20_BASELINE_LIBRARY
-        using namespace test_case_2;
-
-        // P-1
-        ASSERT(
-             bsl::is_void_v<decltype(ntsa::CoroutineUtil::sync_wait(coReturnValue<void>()))>);
-        ASSERT(getValue<int>() == ntsa::CoroutineUtil::sync_wait(coReturnValue<int>()));
-        ASSERT(&getValue<int&>() == &ntsa::SimplTaskUtil::sync_wait(coReturnValue<int&>()));
-        {
-            int&& expectedRef = getValue<int&&>();
-            int&& actualRef   = ntsa::CoroutineUtil::sync_wait(coReturnValue<int&&>());
-            ASSERT(&expectedRef == &actualRef);
-        }
-
-#ifdef BDE_BUILD_TARGET_EXC
-        // P-2
-        {
-            bsl::optional<int> thrownValue;
-
-            auto throwingCoro = []() -> bdxa::CoroutineTask<void> {
-                throw 2;
-            };
-
-            try {
-                ntsa::CoroutineUtil::sync_wait(throwingCoro());
-            } catch (int ex) {
-                thrownValue = ex;
-            }
-
-            ASSERT(2 == thrownValue);
-        }
-#endif
-
-        // P-3
-        struct Awaitable {
-            bsl::jthread d_thread;
-
-            bsl::atomic<bool> d_ready;
-
-            bsl::coroutine_handle<> d_awaiter;
-
-            Awaitable()
-            : d_thread([this] {
-                           while (!d_ready) {
-                               bsl::this_thread::yield();
-                           }
-                           d_awaiter.resume();
-                       })
-            {
-            }
-
-            bool await_ready() { return false; }
-
-            void await_suspend(bsl::coroutine_handle<> awaiter)
-            {
-                d_awaiter = awaiter;
-                d_ready = true;
-            }
-
-            void await_resume() {}
-        } awaitable;
-
-        auto awaitingTask = [&awaitable]() -> bdxa::CoroutineTask<int> {
-            co_await awaitable;
-            co_return 3;
-        }();
-
-        ASSERT(3 == ntsa::CoroutineUtil::sync_wait(bsl::move(awaitingTask)));
-
-        // P-4
-        doP4();
-#else
-        cout << "Skipping test case 2 before C++20...\n";
-#endif
-      } break;
-      case 1: {
-        // --------------------------------------------------------------------
-        // BREATHING TEST
-        //   This case exercises (but does not fully test) basic functionality.
-        //
-        // Concerns:
-        // 1. The class is sufficiently functional to enable comprehensive
-        //    testing in subsequent test cases.
-        // 2. The default template argument for `CoroutineTask` is `void`.
-        //
-        // Plan:
-        // 1. Create and call a `CoroutineTask<>` coroutine.  Verify that the
-        //    return type is `CoroutineTask<void>`, then call `sync_wait` on that
-        //    object.  (C-1)
-        //
-        // Testing:
-        //   BREATHING TEST
-        //   CoroutineTask<>
-        // --------------------------------------------------------------------
-
-        if (verbose) cout << "\nBREATHING TEST"
-                             "\n==============\n";
-
-#ifdef BSLS_LIBRARYFEATURES_HAS_CPP20_BASELINE_LIBRARY
-        auto task = []() -> bdxa::CoroutineTask<> { co_return; }();
-        ASSERT((bsl::is_same_v<bdxa::CoroutineTask<void>, decltype(task)>));
-        ntsa::CoroutineUtil::sync_wait(bsl::move(task));
-#else
-        cout << "Skipping breathing test before C++20...\n";
-        ASSERT(true);
-#endif
-      } break;
-      default: {
-        cerr << "WARNING: CASE `" << test << "' NOT FOUND." "\n";
-        testStatus = -1;
-      }
-    }
-
-    if (testStatus > 0) {
-        cerr << "Error, non-zero test status = " << testStatus << "." "\n";
-    }
-    return testStatus;
-}
-
-// ----------------------------------------------------------------------------
-// NOTICE:
-//      Copyright 2025 Bloomberg Finance L.P. All rights reserved.
-//      Property of Bloomberg Finance L.P. (BFLP)
-//      This software is made available solely pursuant to the
-//      terms of a BFLP license agreement which governs its use.
-// ------------------------------- END-OF-FILE --------------------------------
-
-#endif
